Stop using open-uri. Move public-hosts related methods out of the versions controller.
